Create README.md
Being a programmer is about mastering programming languages, understanding basic concepts in software development, and having the ability to solve problems creatively. It involves continuous learning, collaborating with teammates, and keeping up to date with technology. As a programmer, you will design, develop and test software for various applications and platforms. The ability to adapt quickly to new technologies and possess strong analytical skills are also indispensable.